Improve tests of Literals
